+++ app/src/components/button/styles.tsx
@@ -6,6 +6,7 @@ export const buttonCSS = css`
   font-size: var(--ac-global-dimension-static-font-size-100);
   line-height: 20px; // TODO(mikeldking): move this into a consistent variable
   margin: 0;
+  flex: none;
 
   display: flex;
   gap: var(--ac-global-dimension-static-size-100);
