+++ app/src/pages/prompts/PromptsPage.tsx
@@ -2,14 +2,7 @@ import { useState } from "react";
 import { useLoaderData } from "react-router";
 import invariant from "tiny-invariant";
 
-import {
-  Flex,
-  Heading,
-  Icon,
-  Icons,
-  LinkButton,
-  View,
-} from "@phoenix/components";
+import { Flex, Icon, Icons, LinkButton, View } from "@phoenix/components";
 import { promptsLoader } from "@phoenix/pages/prompts/promptsLoader";
 
 import { PromptsSearch } from "./PromptsSearch";
@@ -23,29 +16,6 @@ export function PromptsPage() {
   return (
     <Flex direction="column" height="100%">
       <View
-        padding="size-200"
-        borderBottomWidth="thin"
-        borderBottomColor="dark"
-        flex="none"
-      >
-        <Flex
-          direction="row"
-          justifyContent="space-between"
-          alignItems="center"
-          gap="size-100"
-        >
-          <Heading level={1}>Prompts</Heading>
-          <LinkButton
-            size="M"
-            leadingVisual={<Icon svg={<Icons.MessageSquareOutline />} />}
-            variant="primary"
-            to="/playground"
-          >
-            Create Prompt
-          </LinkButton>
-        </Flex>
-      </View>
-      <View
         paddingStart="size-200"
         paddingEnd="size-200"
         paddingTop="size-100"
@@ -54,7 +24,22 @@ export function PromptsPage() {
         borderBottomColor="grey-200"
         flex="none"
       >
-        <PromptsSearch onChange={setSearchFilter} />
+        <Flex
+          direction="row"
+          justifyContent="space-between"
+          alignItems="center"
+          gap="size-100"
+        >
+          <PromptsSearch onChange={setSearchFilter} />
+          <LinkButton
+            size="S"
+            leadingVisual={<Icon svg={<Icons.MessageSquareOutline />} />}
+            variant="primary"
+            to="/playground"
+          >
+            New Prompt
+          </LinkButton>
+        </Flex>
       </View>
       <PromptsTable query={loaderData} searchFilter={searchFilter} />
     </Flex>
